backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlmodel>=0.0.22
pydantic>=2.10.3
pydantic-extra-types>=2.9.0
python-multipart>=0.0.20
python-dotenv>=1.0.1
asyncpg>=0.30.0
psycopg2-binary>=2.9.1
slowapi>=0.1.9
httpx>=0.28.0
starlette>=0.37.2,<0.42.0
aiosqlite>=0.20.0
openai>=1.0.0
openai-agents>=0.6.9
cohere>=5.20.1
mcp>=1.11.0
redis>=5.2.0
